fix(layernorm): stop bf16 at the source, not just in the declaration
The registry saying fp32 keeps the BUILD out of bf16; it does not stop a
narrowing inside a kernel or a pin inside a driver, and both were there to be
found.

layernorm/triton/main.py saved x for the backward as `x_2d.to(torch.bfloat16)`,
halving the activation. Every load in these files already widens with
`.to(tl.float32)` and mean/rstd are stored fp32, so that one cast was the
narrowest precision in the whole normalisation, on the one value the backward
cannot recover. It is also a no-op when x really is bf16, so removing it costs
nothing there.

The new test holds all three places bf16 could come back: a `layernorm_*` row
declaring it, a layernorm driver pinning `torch.bfloat16` (as opposed to naming
`drivers.BF16`, which is the switchable name and builds fp32 for an fp32 row),
and a layernorm kernel narrowing a tensor with `.to(torch.bfloat16)` or
`.bfloat16()`. A `dtype=torch.bfloat16` annotation is not matched -- the cute
GEMM's second weight is someone else's operand and says nothing about the
normalisation.

What this does not claim is that the stream is fp32. It is not: token and pair
activations are bf16 and stay bf16. The normalisation over them is what has to
be wide.

+++ src/miniworld_engine/kernels/layernorm/triton/main.py
@@ -316,7 +316,14 @@ class TritonLayerNormFunction(torch.autograd.Function):
         )
 
         ctx.save_for_backward(
-            x_2d.to(torch.bfloat16),
+            # x at its OWN dtype. This used to be `.to(torch.bfloat16)`, halving the saved
+            # activation -- and a layernorm backward recomputed from a bf16 copy of x is the
+            # narrowest precision in the whole normalisation. Every load in these kernels already
+            # widens with `.to(tl.float32)` and mean/rstd are stored fp32; saving x narrow put the
+            # one value that cannot be recovered back at bf16. The cast is also a no-op when x
+            # really is bf16, so removing it costs nothing there and keeps fp32 where it is asked
+            # for.
+            x_2d,
             weight,
             mean,
             rstd,
